further refinements to api, particularly related to timeouts
git-svn-id: https://svn.osgeo.org/gdal@19086 f0d54148-0727-0410-94bb-9a71ac55c965

+++ sandbox/normanb/gdal/frmts/jpipkak/jpipkakdataset.h
@@ -245,7 +245,7 @@ public:
     JPIPKAKAsyncReader();
     virtual ~JPIPKAKAsyncReader();
 
-    virtual GDALAsyncStatusType GetNextUpdatedRegion(int timeout,
+    virtual GDALAsyncStatusType GetNextUpdatedRegion(double timeout,
                                                      int* pnxbufoff,
                                                      int* pnybufoff,
                                                      int* pnxbufsize,
